requirements.txt
argparse
networkx>=1.10
matplotlib>=1.5.0
mysql.connector>=2.1.3
